+++ Source/DataTransfer.Legacy.Services/AuditService.cs
@@ -1,18 +1,25 @@
 ﻿using System.Threading.Tasks;
+using Castle.Core;
 using Relativity.Core.Service;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1.Models;
 using Relativity.DataTransfer.Legacy.Services.Helpers;
+using Relativity.DataTransfer.Legacy.Services.Interceptors;
 using Relativity.DataTransfer.Legacy.Services.Runners;
 
 namespace Relativity.DataTransfer.Legacy.Services
 {
+	[Interceptor(typeof(PermissionCheckInterceptor))]
+	[Interceptor(typeof(LogInterceptor))]
+	[Interceptor(typeof(MetricsInterceptor))]
+	[Interceptor(typeof(UnhandledExceptionInterceptor))]
 	public class AuditService : BaseService, IAuditService
 	{
 		private readonly IMassExportManager _massExportManager;
 		private readonly MassImportManager _massImportManager;
 
-		public AuditService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) : base(methodRunner, serviceContextFactory)
+		public AuditService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) 
+			: base(methodRunner, serviceContextFactory)
 		{
 			_massExportManager = new MassExportManager();
 			_massImportManager = new MassImportManager();