+++ Source/DataTransfer.Legacy.Services/FileIOService.cs
@@ -1,19 +1,27 @@
 ﻿using System.Threading.Tasks;
+using Castle.Core;
 using Relativity.Core;
 using Relativity.Core.DTO;
 using Relativity.Core.Service;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1.Models;
 using Relativity.DataTransfer.Legacy.Services.Helpers;
+using Relativity.DataTransfer.Legacy.Services.Interceptors;
 using Relativity.DataTransfer.Legacy.Services.Runners;
 
 namespace Relativity.DataTransfer.Legacy.Services
 {
+	[Interceptor(typeof(PermissionCheckInterceptor))]
+	[Interceptor(typeof(LogInterceptor))]
+	[Interceptor(typeof(MetricsInterceptor))]
+	[Interceptor(typeof(UnhandledExceptionInterceptor))]
+
 	public class FileIOService : BaseService, IFileIOService
 	{
 		private readonly ExternalIO _externalIo;
 
-		public FileIOService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) : base(methodRunner, serviceContextFactory)
+		public FileIOService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) 
+			: base(methodRunner, serviceContextFactory)
 		{
 			_externalIo = new ExternalIO();
 		}
@@ -43,9 +51,8 @@ namespace Relativity.DataTransfer.Legacy.Services
 		{
 			return ExecuteAsync(() =>
 				{
-					var serviceContext = GetBaseServiceContext(AdminWorkspace);
-					string documentDirectory = GetDefaultDocumentDirectory(serviceContext, workspaceID);
-					_externalIo.ExternalRemoveFill(serviceContext, documentDirectory, fileName, workspaceID);
+					string documentDirectory = GetDefaultDocumentDirectory(workspaceID);
+					_externalIo.ExternalRemoveFill(GetBaseServiceContext(workspaceID), documentDirectory, fileName, workspaceID);
 				},
 				workspaceID, correlationID);
 		}
@@ -54,8 +61,7 @@ namespace Relativity.DataTransfer.Legacy.Services
 		{
 			return ExecuteAsync(() =>
 			{
-				var serviceContext = GetBaseServiceContext(AdminWorkspace);
-				string documentDirectory = GetDefaultDocumentDirectory(serviceContext, workspaceID);
+				string documentDirectory = GetDefaultDocumentDirectory(workspaceID);
 				return _externalIo.GetRepositoryReport(documentDirectory);
 			}, workspaceID, correlationID);
 		}
@@ -82,8 +88,9 @@ namespace Relativity.DataTransfer.Legacy.Services
 			return ExecuteAsync(() => Config.RepositoryVolumeMax, null, correlationID);
 		}
 
-		private string GetDefaultDocumentDirectory(BaseServiceContext serviceContext, int workspaceID)
+		private string GetDefaultDocumentDirectory(int workspaceID)
 		{
+			BaseServiceContext serviceContext = GetBaseServiceContext(AdminWorkspace);
 			CaseManager caseManager = new CaseManager();
 			int fileLocationCodeArtifactID = caseManager.Read(serviceContext, workspaceID).DefaultFileLocationCodeArtifactID;
 			ResourceServer resourceServer = ResourceServerManager.Read(serviceContext, fileLocationCodeArtifactID);