+++ Source/DataTransfer.Legacy.Services/Interceptors/RelativityPermissionHelper.cs
@@ -1,0 +1,12 @@
+﻿using Relativity.Core;
+
+namespace Relativity.DataTransfer.Legacy.Services.Interceptors
+{
+	public class RelativityPermissionHelper : IRelativityPermissionHelper
+	{
+		public bool HasAdminOperationPermission(ICoreContext coreContext, Permission permission)
+		{
+			return PermissionsHelper.HasAdminOperationPermission(coreContext, permission);
+		}
+	}
+}