+++ Source/DataTransfer.Legacy.Services/ProductionService.cs
@@ -2,30 +2,38 @@
 using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
+using Castle.Core;
 using Relativity.Core.Service;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1;
 using Relativity.DataTransfer.Legacy.SDK.ImportExport.V1.Models;
 using Relativity.DataTransfer.Legacy.Services.Helpers;
+using Relativity.DataTransfer.Legacy.Services.Interceptors;
 using Relativity.DataTransfer.Legacy.Services.Runners;
 using Relativity.Export;
 
 namespace Relativity.DataTransfer.Legacy.Services
 {
+	[Interceptor(typeof(PermissionCheckInterceptor))]
+	[Interceptor(typeof(LogInterceptor))]
+	[Interceptor(typeof(MetricsInterceptor))]
+	[Interceptor(typeof(UnhandledExceptionInterceptor))]
 	public class ProductionService : BaseService, IProductionService
 	{
 		private readonly ProductionManager _productionManager;
 
-		public ProductionService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) : base(methodRunner, serviceContextFactory)
+		public ProductionService(IMethodRunner methodRunner, IServiceContextFactory serviceContextFactory) 
+			: base(methodRunner, serviceContextFactory)
 		{
 			_productionManager = new ProductionManager();
 		}
 
-		public Task<object[][]> RetrieveBatesByProductionAndDocumentAsync(int workspaceID, int[] productionIDs, int[] documentIDs, string correlationID)
+		public Task<ExportDataWrapper> RetrieveBatesByProductionAndDocumentAsync(int workspaceID, int[] productionIDs, int[] documentIDs, string correlationID)
 		{
 			return ExecuteAsync(() =>
 			{
-				var result = ProductionQuery.RetrieveBatesByProductionAndDocument(GetBaseServiceContext(workspaceID), GetUserAclMatrix(workspaceID), productionIDs, documentIDs);
-                return ToObjectArrays(result, ProductionDocumentBatesHelper.ToSerializableObjectArray);
+				kCura.Data.DataView resultAsDataView = ProductionQuery.RetrieveBatesByProductionAndDocument(GetBaseServiceContext(workspaceID), GetUserAclMatrix(workspaceID), productionIDs, documentIDs);
+				object[][] resultAsObjectArrays =  ToObjectArrays(resultAsDataView, ProductionDocumentBatesHelper.ToSerializableObjectArray);
+				return new ExportDataWrapper(resultAsObjectArrays);
 			}, workspaceID, correlationID);
 		}